+++ src/dao/AccountDAO.js
@@ -1,25 +1,9 @@
 // @flow
+import type { Account } from '../types/entities/Account';
+
 import DAO from './DAO';
 import { DAO_ENTITIES } from '../constants';
 import DefaultTranslator from '../translators/DefaultTranslator';
-
-export type Account = {
-  accessFailedCount: number,
-  banned: boolean,
-  createdDate: string,
-  email: string,
-  emailConfirmed: boolean,
-  fullName: ?string,
-  id: string,
-  lockoutEnabled: boolean,
-  lockoutEndDateUtc: ?string,
-  logins: Object,
-  phoneNumber: ?number,
-  phoneNumberConfirmed: boolean,
-  roles: Object,
-  twoFactorEnabled: boolean,
-  userName: string,
-};
 
 class AccountDAO extends DAO<Account, Account> {
   constructor() {
