+++ src/types/entities/Availability.js
@@ -1,0 +1,6 @@
+// @flow
+export type Availability = {
+  description: ?string,
+  id: string,
+  name: string,
+};
