+++ src/types/entities/Device.js
@@ -1,0 +1,19 @@
+// @flow
+export type Device = {
+  createdBy: {
+    id: string,
+    userName: string,
+  },
+  deviceStatus: string,
+  id: number,
+  isDeleted: boolean,
+  isMultiTap: boolean,
+  lastEdited: string,
+  lastEditedBy: {
+    id: string,
+    userName: string,
+  },
+  name: string,
+  particleId: string,
+  temperature: number,
+};
