+++ src/types/entities/Glass.js
@@ -1,0 +1,8 @@
+// @flow
+export type Glass = {
+  createDate: Date,
+  description: ?string,
+  id: string,
+  name: string,
+  updateDate: Date,
+};
