+++ src/types/entities/Location.js
@@ -1,0 +1,16 @@
+// @flow
+export type Location = {
+  city: string,
+  createdDate: string,
+  description: string,
+  geolocation: Object,
+  id: number,
+  locationType: string,
+  name: string,
+  state: ?string,
+  street: string,
+  suite: string,
+  summary: string,
+  timeZone: string,
+  zipCode: number,
+};
