+++ src/types/entities/Srm.js
@@ -1,0 +1,6 @@
+// @flow
+export type Srm = {
+  hex: string,
+  id: string,
+  name: string,
+};
