+++ src/types/entities/Style.js
@@ -1,0 +1,18 @@
+// @flow
+export type Style = {
+  abvMax: number,
+  abvMin: number,
+  createDate: Date,
+  description: ?string,
+  fgMax: number,
+  fgMin: number,
+  ibuMax: number,
+  ibuMin: number,
+  id: string,
+  name: string,
+  ogMax: number,
+  ogMin: number,
+  srmMax: number,
+  srmMin: number,
+  updateDate: Date,
+};
